Re-organizes OrderedSet
