+++ src/schema/v2/schema.ts
@@ -35,7 +35,7 @@ import { CreditCard } from "./credit_card"
 // import Fairs from "./fairs"
 import Genes from "./genes"
 import GeneFamilies from "./gene_families"
-import OrderedSet from "./ordered_set"
+import { OrderedSet } from "./OrderedSet"
 import OrderedSets from "./ordered_sets"
 // import Profile from "./profile"
 // import Partner from "./partner"
